Move to C++20 (#113)
* Move to C++20

* std::enable_if to requires

* Updated build tools

* clang compilation fix

* Update build tools (for SDL compilation)

* GCC compilation fixes

* clang 18 doesn't has zoned_time, return workaround

* Mac compilation fix

* Backward update

* Missed FO_STACK_TRACE_ENTRY

* iOS sdk 13 -> 18

* AnyData float formatting fix

* Exclude libunwind for mac compilation fix

* Formatting fix

* WinApi includes fix

* Missed include

* Use std::ranges

* Build fixes

* Build fixes

* Compilation fixes

* uint -> uint32

* int -> int32

* float -> float32, double -> float64

* Arithmetic conversion from static_cast to numeric_cast

* Simplifying of int/uint usage, move mostly to int

* GCC compilation fix

* Renaming

* Common module split to smaller Essentials modules

* Compilation fix

* Compilation fix

* Refactroing

* Compilation fix

* Compilation fix

* lerp fix

+++ Source/Applications/ServerDaemonApp.cpp
@@ -34,7 +34,6 @@
 #include "Common.h"
 
 #include "Application.h"
-#include "Platform.h"
 #include "Server.h"
 #include "Settings.h"
 
@@ -51,7 +50,7 @@ int main(int argc, char** argv)
     try {
         Platform::ForkProcess();
 
-        InitApp(argc, argv);
+        InitApp(numeric_cast<int32>(argc), argv);
 
         {
             auto server = SafeAlloc::MakeRefCounted<FOServer>(App->Settings);
